extract starter code to Rails Controller Concern
