+++ app/controllers/concerns/starter_code.rb
@@ -1,0 +1,8 @@
+module StarterCode
+  def starter_code_repository_id(repo_name)
+    return unless repo_name.present?
+    sanitized_repo_name = repo_name.gsub(/\s+/, '')
+    github_repository   = GitHubRepository.new(current_user.github_client, nil)
+    github_repository.repository(sanitized_repo_name).id
+  end
+end
